requirements.txt
wheel
PyOpenGL
wxPython  # apt install libgtk-3-dev
networkx
numpy
spglib
pyyaml
jinja2
